+++ ChangeLog.md
@@ -1,5 +1,19 @@
 # Change Log
 
+### 2022-03-XX - 5.14.1
+
+| Library                                                                 | Min Version |
+| ----------------------------------------------------------------------- | ----------- |
+| `@autorest/core`                                                        | `3.6.2`     |
+| `@autorest/modelerfour`                                                 | `4.19.1`    |
+| `azure-core` dep of generated code                                      | `1.20.1`    |
+| `msrest` dep of generated code                                          | `0.6.21`    |
+| `azure-mgmt-core` dep of generated code (If generating mgmt plane code) | `1.3.0`     |
+
+**New Features**
+
+- Support flag `--security`/`--security-scopes`/`--security-header-name` and security definition in swagger #1161
+
 ### 2022-03-08 - 5.14.0
 
 | Library                                                                 | Min Version |